it looks like passing enums to va_start is unsafe as codechecker reports "passing an object that undergoes default argument promotion to 'va_start' has undefined behavior". Maybe adding a casting to int (forcing the int conversion) will fix it.

+++ core/Float.cc
@@ -403,7 +403,7 @@ void FLOAT::decode(const TTCN_Typedescriptor_t& p_td, TTCN_Buffer& p_buf,
                    TTCN_EncDec::coding_t p_coding, ...)
 {
   va_list pvar;
-  va_start(pvar, p_coding);
+  va_start(pvar, (int)p_coding);
   switch(p_coding) {
   case TTCN_EncDec::CT_BER: {
     TTCN_EncDec_ErrorContext ec("While BER-decoding type '%s': ", p_td.name);
